requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv